Rewrite of stage 1 boss
The original code was really fucking horrible to look at despite my best
efforts at commenting it. The boss was forcing variables to orbs and the
orbs were forcing it back leading to a lot of back-and-forth debugging
to figure out if anything went wrong in either of the three objects, now
it's a one-way street in that regard and the orbs should work by themselves.
Also a balance adjustment on when health-based attacks by orb2 occur.