Partial fix to the stage 2 problem
It works but I'm concerned it's not effective in the long term for good
placing of the initial view screen. Problem is:
- adjustment of all rooms by scr_screen_init() resets view defaults set
in Room Editor
- Creation of objects and their creation code occurs before Room Creation and Room Start
- Room Start is where the controller asserts view position

Changing said event to Create event results in variable errors because said
vars are either called in object Creation code or Room Creation code.
Perhaps changing instance order in Room Editor might work?