Laser improvements
Took a bit of draw_sprite_general wrangling
but I was able to get the head to move smoothly
not only along the laser point but also
drawn partially when just coming out from
the source. Head x offset is slightly further
from edge to cover the laser end properly.